🔨 Refactor: 하루 건강 데이터 전송 API 연결 및 제출 UI 수정

+++ WAYVI/WAYVI Watch App/Views/HealthSubmitPromptContentView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct HealthSubmitPromptContentView: View {
-    let userId: Int64
+    let userId: Int
     let healthData: HealthData
     var onComplete: () -> Void
